Moved BlockFetcherIterator to its own file.
